Cluster contextualization
User may optionally provide additional information in the 'ucernvm'
section of the context file to use a master/slave cluster
contextualization.

Master sends its IP address to the central service (cernvm-online.cern.ch),
from where it is fetched by the slave machines. The slave machines
are waiting for the master to publish the data (blocking wait in the
bootloader).

This is the cloud-init part of the contextualization process.

First service sends the 'master_ready' status to the service, letting
the slave machines know they can fetch the data.

Second service does the master data fetching and placeholder replacement
for cloud-init context file (which cannot be done by amiconfig).